Removing workflow concurrency block
The concurrency was set at the workflow level which was probably wrong from the start and it used a fixed group name. What that meant was that a PR run would cancel a run that was triggered by pushing master.

Since PR builds don't push, it is OK to have them run anytime. And since other builds will push to a tag derived from the current hash it is fine for them to run concurrently.

If you really need to make sure workflows or jobs in workflows are not running concurrently, this is the only tool you have and it is not good because YOUR ONLY OPTION is to cancel the operation that is currently running. There is no way to wait for a workflow/job that is currently running before allowing a following operation proceed.

https://docs.github.com/en/actions/how-tos/write-workflows/choose-when-workflows-run/control-workflow-concurrency